# Optimize ISO datetime parsing using binary pattern matching with nibbles