@dkuku 2025-01-11
I know, this is a good candidate for a macro.