@josevalim 2025-01-22
@dkuku why don't we give such a macro a try then on this change?

```
defmacrop ascii_char(x) do
  quote do
    <<3::4, unquote(x)::4>>
  end
end
```

And then it may clean up the code a bit?