@dkuku 2025-01-22
@josevalim  It works but I need to define it in a separate module because - probably because this code is not inside a function.
Also ascii does not fit here, maybe int fits better (you can compare both commits.)