@dkuku 2025-01-22
The longer I think about it the more I think such match should be implemented in otp. Matching and parsing integers from binaries is much more common than utf.